+++ sdk/@launchdarkly/flutter/packages/observability/lib/launchdarkly_flutter_observability.dart
@@ -1,6 +1,12 @@
 export 'src/ld_observe.dart' show LDObserve;
+export 'src/instrumentation/navigation/ld_navigator_observer.dart'
+    show LDNavigatorObserver, LDScreenNameExtractor;
+export 'src/instrumentation/navigation/route_patterns.dart'
+    show LDRoutePatterns;
 export 'src/session_replay_capture.dart' show SessionReplayCapture;
 export 'src/masking.dart' show LDMask, LDUnmask, LDIgnore;
+export 'src/instrumentation/click/ld_click.dart'
+    show LDClick, LDClickTargetInfo, LDClickTargetResolver;
 export 'src/options/observability_options.dart'
     show
         ObservabilityOptions,